Rename test files so harmonize with Kotlin version and improve understanding
